requirements.txt
fastapi
uvicorn
beautifulsoup4
tldextract
pillow
pytesseract
pyzbar
imagehash
python-magic
exifread
transformers
torch
scikit-learn
lightgbm
numpy
pandas
lxml
joblib
matplotlib
python-dotenv
loguru
